installed prettier and configured vs code workspace
VS Code doesn't support folder-specific settings. Therefore, we need a multi-root workspace. The Workspace is now configured in a way, that Deno is enabled with Deno formatter and linter only in the backend folder. All other folders use the locally installed prettier as the default formatter (on save). Deno config needs adjustment for formatting and linting! ESLint and commit hooks still need to be set up and integrated with prettier for the react app!