removed deno.json to get recommended default configuration
given deno.json was an example from documentation.

As Deno says: "Note that using a configuration file is not required now, and will not be required in the future. Deno still works best with the default options and no configuration file. [...] Using the configuration file should be considered an 'as needed' feature, not something every user should be reaching to as the first thing when setting up a project."